requirements.txt
matplotlib == 3.4.3
nest_asyncio == 1.5.1
numpy == 1.19.5
pandas == 1.3.2
requests == 2.26.0
scikit_learn == 0.24.2
seaborn == 0.11.2
spacy == 3.3.1
streamlit == 1.9.2
streamlit_chat == 0.0.2.1
tweet_preprocessor == 0.6.0
twint == 2.1.21
matplotlib==3.4.3
nltk==3.7
tweet-preprocessor==0.6.0
tensorflow==2.8.0
flask==2.1.2

